changed "permutations.rs" to "permutation.rs" and added module-level document comment
